requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
langchain==0.1.15
langchain-openai==0.1.0
requests==2.31.0
beautifulsoup4==4.12.3
python-dotenv==1.0.1
flask==3.0.3
flask-cors==4.0.1
flask-socketio==5.3.6
python-socketio==5.11.2
python-engineio==4.8.2
uuid==1.30
